fix(ui-touch): workaround the fact that we miss touch events
The touch controller generates pulses for the events.
It doesn't assert the line until we the MCU reads the data out.
When the UI is plotting, currently we block the MCU and this can cause to miss some interrupts (since we await for the event instead of doing actual work in interrupt context).