perf(display): avoid using the buffered version of the display driver
Since slint does already the job of updating only the necessary pixels & lines, we don't need to have a buffer of the whole display in the driver as well.
Even with the double-buffer approach, again Slint would want to have access to the buffer itself hence tehre is no need to have an additional hidden buffer.